requirements.txt
# Core dependencies
python-docx==1.1.0
PyMuPDF==1.24.0
python-dotenv==1.0.0

# LLM clients (install the one you need)
openai>=1.0.0
anthropic>=0.18.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0

# Optional utilities
aiohttp>=3.9.0
pydantic>=2.0.0
